fix(guitarpro): parse more real-world GP3/5/7 files
Real-world corpus sweep: 244 files, 10 failing → 4 failing. Fixes cover
three independent bugs plus a build-blocker from a pre-commit typo fixer
that only rewrote half the occurrences of a misspelled field.

* GP5: `read_duration` no longer panics on out-of-range duration bytes
  (compute the shift in i16 before the range check).
* GP7 .gp: `<Automation><Position>` is a fractional bar position (0.0-1.0),
  so `Automation.position` is deserialised as f32 instead of i32.
* CLI loader: detect the version from the "FICHIER GUITAR PRO vX.YZ" magic
  instead of the on-disk extension, so mis-labelled files (a real v5.10
  file with a .gp3 extension) load correctly, and non-GP content produces
  a readable error.
* rename `show_rythm`/`extend_rythmic` -> `show_rhythm`/`extend_rhythmic`
  everywhere (the typos hook had rewritten only half the occurrences).

+++ guitarpro/src/io/gpif.rs
@@ -82,8 +82,10 @@ pub struct Automation {
     pub value: String,
     #[serde(rename = "Bar", default)]
     pub bar: i32,
+    // GP7 stores this as a fractional position within the bar in `[0.0, 1.0)`
+    // (e.g. `0.75`), not an integer tick offset.
     #[serde(rename = "Position", default)]
-    pub position: i32,
+    pub position: f32,
 }
 
 // ---------------------------------------------------------------------------
